pyproject.toml
[tool.poetry]
name = "eile_forecast"
version = "0.0.1"
description = "Load forecasting benchmarks for energy management systems in SMEs."
authors = ["Claudia Ehrig <claudia.ehrig@iis.fraunhofer.de>"]
readme = "README.md"
license = "GPL-3.0"
packages = [{include = "eile_forecast", from = "src"}]

[tool.poetry.dependencies]
python = "3.10.18"
hydra-core = "1.3.0"
matplotlib = "3.8.0"
numpy = "1.26.4"
omegaconf = "2.3.0"
pandas = "2.0.0"
pyarrow = "12.0.0"
xgboost ="2.0.3"
onnxmltools ="1.0.0.0"
skl2onnx ="1.16.0"
holidays = "0.47.0"
loguru = "0.7.2"
torch={version="2.3.1+cpu",source="pytorch"}
plotly = "5.24.1"
scikit-learn = "1.4.2"
gluonts = "0.13"
utilsforecast = "0.2.12"
statsforecast = "2.0.2"
seaborn = "0.13.2"
timezonefinder = "6.5.9"
requests = "2.32.4"
keras = "3.11.2"
tensorflow = "2.20.0"

[[tool.poetry.source]]
name="pytorch"
url="https://download.pytorch.org/whl/cpu"
priority="explicit"


[tool.poetry.group.dev.dependencies]
ipykernel = "6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120

[tool.mypy]
ignore_missing_imports = true
show_column_numbers = true

[tool.pylint]
disable = "C0111"
good-names-rgxs = "^[_a-z][_a-z0-9]?$"
max-line-length = 120
